Handle some errors in SimplePing.
Do not fail directly on EHOSTUNREACH or DNS errors.
Instead retry these when the next PING is due.

Detect ENOMEM which means "out of socket descriptors" or something.
It's not clear why it happens, all sockets should be closed properly.
Let the caller handle this, might need even a machine.reset() to fix.